MNT: Remove bitstring dependency
This removes the bitstring dependency and all bitstring inputs
allowed with it. You can pass in the raw objects instead of the
bitstring objects now.

All internal bookkeeping is done with a PacketData class that
is just a thin wrapper around bytes data to keep track of the
amount of data read.