MNT: Use built-in types for return data objects
This adds the ability to subclass the built-in types and return an
item that can be acted upon directly rather than having to access
`parsed_item.raw_value` as an attribute to do any work with the item.

The built-in types are immutable and can't have attributes added to
them so we inject the `raw_value` attribute in the new constructor
of each explicit type.