MNT: Change to returning intermediate values for enumerated and boolean parameters
This gives the ability to get back to the raw value if desired still, but returns
what was used for the "lookup" / bool conversion as that is likely what
a user needs for the first debug step. Then they can get to the raw-raw
value themselves if they really need it.